authJWT inclosa

+++ server/controllers/apiUser.js
@@ -129,6 +129,11 @@ exports.loginUser = async (req, res) => {
 exports.deleteUser = async (req, res) => {
   let usr = await User.findById({_id: req.params.id})
   //let email = req.params.email; 
+
+  if (usr.userId == req.user.id) {
+    res.status(401).json({error: "Do not have permission"})
+    return;
+  }
   try{
     let usr = await User.findById({_id: req.params.id})
     usr.delete();
@@ -146,8 +151,9 @@ exports.updateUser = async (req, res) => {
     const ecoPoints = req.body.ecoPoints;
     const score = req.body.score;
   
-    const id = req.params.id;
+    const id = req.user.id;
     const user = await User.findById(id)
+
     console.log("Searching for user to update: " + req.params.id);
 
     if (level != null)  user.level = level;
@@ -183,7 +189,7 @@ exports.getUserProducts = async (req, res) => {
 
 exports.rateUser = async (req, res) => {
   // Id usuario a valorar
-  const userId = req.params.userId;
+  const userId = req.params.id;
   const rateScore = req.body.rateScore;
   const comment = req.body.comment;
 
@@ -274,7 +280,7 @@ exports.getMyCommentsRecived = async (req, res) => {
 
 exports.getRewards = async (req, res) => {
   try {
-    const user = await User.findById({ _id: req.params.id });
+    const user = await User.findById({ _id: req.user.id });
     var ngifts = user.gifts;
     var nloans = user.loans;
     var nexchanges = user.exchanges;
@@ -324,7 +330,7 @@ exports.getUserLevel = async (req, res) => {
 
 exports.levelManage = async (req, res) => {
   try {
-    const user = await User.findById({ _id: req.params.id });
+    const user = await User.findById({ _id: req.user.id });
     console.log("Level del usuario: " , user.name);
 
     var nivelAntiguo = user.level;
@@ -372,7 +378,7 @@ exports.getUserRewards = async (req, res) => {
   try {
     const {type, estimatedPoints} = req.body;
 
-    const id = req.params.id;
+    const id = req.user.id;
     const user = await User.findById(id)
     console.log("Searching for user to get reward: " + user.name);
     
@@ -389,7 +395,7 @@ exports.getUserRewards = async (req, res) => {
 
 exports.getUserWishlist = async (req, res) => {
   try {
-    const userId = req.params.id;
+    const userId = req.user.id;
     User.findOne({_id: userId}, (erro, user) => {
       console.log(user.wishlist)
       console.log(user);
@@ -404,7 +410,7 @@ exports.getUserWishlist = async (req, res) => {
     
 exports.addToWishlist = async (req, res) => {
   try {
-    const userId = req.params.id;
+    const userId = req.user.id;
     const ourUser = await User.findById({_id: userId});
     let idProduct = req.body.idProduct;
     
@@ -426,7 +432,7 @@ exports.addToWishlist = async (req, res) => {
 
 exports.deleteFromWishlist = async (req, res) => {
   try {
-    const userId = req.params.id;
+    const userId = req.user.id;
     let idProduct = req.body.idProduct;
     User.findById({_id: userId}, {wishlist: 1}, async (erro, usersProducts) => {
         let find = 0;
@@ -477,7 +483,7 @@ exports.getUserFollowers = async (req, res) => {
 
 exports.follow = async (req, res) => {
   try {
-    const userId = req.params.id;
+    const userId = req.user.id;
     let body = req.body;
     const ourUser = await User.findById({_id: userId});
     const userFollowed = await User.findOne ({email: body.email});
